+++ web/typescript/frontend/InitiatorForm.ts
@@ -1,3 +1,4 @@
+// noinspection JSUnusedLocalSymbols
 const CONTACT_NONE = 0;
 const CONTACT_OPTIONAL = 1;
 const CONTACT_REQUIRED = 2;